renamed sprite to drawable_object
