+++ source/drawable_objects/cell/cell.cpp
