+++ source/drawable_objects/cell/coord_translation.h
