+++ source/drawable_objects/cell/hexagon.cpp
@@ -1,6 +1,6 @@
-#include <source/sprites/cell/hexagon.h>
+#include <source/drawable_objects/cell/hexagon.h>
 #include <source/options/options.h>
-#include <source/sprites/cell/cell.h>
+#include <source/drawable_objects/cell/cell.h>
 
 Hexagon::Hexagon(const Cell& cell) : cell_(cell) {}
