+++ source/drawable_objects/sprite.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-class Sprite {
+class DrawableObject {
 public:
     //virtual void HandleEvent(const sf::Event&) = 0;
     virtual void Draw(Screen&, const GameOptions&) = 0;
