+++ source/drawable_objects_groups/game_scene/game_scene.h
@@ -1,5 +1,5 @@
-#include <source/sprites_groups/scene.h>
-#include <source/sprites_groups/game_scene/grid.h>
+#include <source/drawable_objects_groups/scene.h>
+#include <source/drawable_objects_groups/game_scene/grid.h>
 #include <source/utility/all_usings.h>
 
 #pragma once
