+++ source/drawable_objects_groups/game_scene/grid.cpp
@@ -1,5 +1,5 @@
-#include <source/sprites_groups/game_scene/grid.h>
-#include <source/sprites/cell/cell.h>
+#include <source/drawable_objects_groups/game_scene/grid.h>
+#include <source/drawable_objects/cell/cell.h>
 #include <tuple>
 
 Grid::Grid(const GameOptions& game_options) {
@@ -7,7 +7,7 @@ Grid::Grid(const GameOptions& game_options) {
     for (size_t i = 0; i < 20; ++i) {
         for (size_t j = 0; j < 10; ++j) {
             cells_[i].push_back(make_unique<Cell>(std::make_pair(i + 3, j + 1), game_options));
-            sprites_.push_back(cells_[i][j].get());
+            drawable_objects_.push_back(cells_[i][j].get());
         }
     }
 }
