+++ source/drawable_objects_groups/game_scene/grid.h
@@ -1,12 +1,12 @@
-#include <source/sprites_groups/sprites_group.h>
+#include <source/drawable_objects_groups/drawable_objects_group.h>
 #include <source/sfml_facades/screen.h>
-#include <source/sprites/cell/cell.h>
+#include <source/drawable_objects/cell/cell.h>
 #include <source/options/options.h>
 #include <vector>
 
 #pragma once
 
-class Grid : public SpritesGroup {
+class Grid : public DrawableObjectsGroup {
     std::vector<std::vector<unique_ptr<Cell>>> cells_;
 public:
     Grid(const GameOptions&);
