+++ source/drawable_objects_groups/scene.cpp
@@ -1,5 +1,5 @@
 #include <source/sfml_facades/screen.h>
-#include <source/sprites_groups/scene.h>
+#include <source/drawable_objects_groups/scene.h>
 
 void Scene::LoopIteration(Screen& screen, const GameOptions& game_options) {
     Draw(screen, game_options);
