+++ source/drawable_objects_groups/scene.h
@@ -1,10 +1,10 @@
-#include <source/sprites_groups/sprites_group.h>
+#include <source/drawable_objects_groups/drawable_objects_group.h>
 #include <source/sfml_facades/screen.h>
 #include <source/options/options.h>
 
 #pragma once
 
-class Scene : public SpritesGroup {
+class Scene : public DrawableObjectsGroup {
 public:
     virtual void LoopIteration(Screen&, const GameOptions&);
 };
